requirements.txt
tk
python-dotenv
psycopg2-binary